fix:TopUserList to page
修正所以頁面的UserList引用方式
