--- src/components/TopUser/TopUser.module.scss
@@ -1,30 +1,0 @@
-.userContainer {
-  display: flex;
-  padding: 16px;
-  border-bottom: 1px solid #e6ecf0;
-}
-.avatar {
-  border-radius: 50%;
-}
-.info {
-  padding-left: 8px;
-  padding-right: 13px;
-}
-.name {
-  font-weight: 700;
-  font-size: 16px;
-  line-height: 26px;
-  &:hover {
-    text-decoration-line: underline;
-  }
-}
-.account {
-  font-weight: 500;
-  font-size: 14px;
-  color: #92929d;
-}
-
-.btn {
-  position: absolute;
-  right: 16px;
-}
