+++ src/components/TopUserList/TopUserList.module.scss
@@ -17,3 +17,31 @@
   padding-bottom: 24px;
   border-bottom: 1px solid #e6ecf0;
 }
+
+.userContainer {
+  display: flex;
+  padding: 16px;
+  border-bottom: 1px solid #e6ecf0;
+}
+.info {
+  padding-left: 8px;
+  padding-right: 13px;
+}
+.name {
+  font-weight: 700;
+  font-size: 16px;
+  line-height: 26px;
+}
+.account {
+  font-weight: 500;
+  font-size: 14px;
+  color: #92929d;
+}
+.btn {
+  position: absolute;
+  right: 16px;
+}
+
+.avatar {
+  border-radius: 50%;
+}
