+++ src/pages/UserProfilePage.jsx
@@ -2,7 +2,7 @@
 import UserSidebar from "../components/UserSidebar/UserSidebar";
 import UserProfilePart from "../components/UserProfilePart/UserProfilePart";
 import UserProfileTabs from "../components/UserProfileTabs/UserProfileTabs";
-import TopUserList from '../components/TopUserList/TopUserList'
+import TopUserList from "../components/TopUserList/TopUserList";
 
 //react-bootstrap
 import Container from "react-bootstrap/Container";
@@ -16,15 +16,13 @@ import { Outlet, useLocation, useNavigate } from "react-router-dom";
 import { useState, useEffect } from "react";
 
 //api
-import { getTopUsers, getUserData } from "../apis/userData";
+import { getUserData } from "../apis/userData";
 import { useAuth } from "../contexts/AuthContext";
 
 const UserProfilePage = () => {
   ///////////update userData////////////
   //userData狀態
   const [userData, setUserData] = useState();
-  const [topUsers, setTopUsers] = useState();
-
 
   const { currentMember } = useAuth();
 
@@ -54,18 +52,6 @@ const UserProfilePage = () => {
     userInfo();
   }, [currentMember.id, urlUserId, navigate]);
 
-  useEffect(() => {
-    const topUsers = async () => {
-      try {
-        const { data } = await getTopUsers();
-        setTopUsers(data);
-      } catch (error) {
-        console.log(error);
-      }
-    };
-    topUsers();
-  }, []);
-
   ///////////update userData////////////
   return (
     <Container>
@@ -82,7 +68,7 @@ const UserProfilePage = () => {
         </Col>
         <Col xs={3}>
           <div className="sticky-top">
-            <TopUserList topUsers={topUsers}/>
+            <TopUserList />
           </div>
         </Col>
       </Row>
